pyproject.toml
[tool.poetry]
name = "yesql"
version = "0.1.0"
description = "YeSQL is a SQL-first data manipulation library that will replace your ORM."
authors = ["Sean Stewart <sean_stewart@me.com>"]
license = "MIT"

[tool.poetry.dependencies]
# Library
python = "^3.8"
typical = "^2.8"
PyPika = "^0.48.0"
asyncpg = {version = "^0.25.0", optional = true}
orjson = {version = "^3.5.1", optional = true}
psycopg = {extras = ["binary", "pool"], version = "^3.0", optional = true}
# Docs
mkdocs = {version = "^1.2.2", optional = true}
mkdocs-material = {version = "^8.0", optional = true}
mkdocs-awesome-pages-plugin = {version = "^2.5.0", optional = true}
# Lint
mypy = {version = "^0.910", optional = true}
black = {version = "^21.7b0", optional = true}
flake8 = {version = "^3.9.2", optional = true}
# Test
pytest = {version = "^6.2.4", optional = true}
pytest-asyncio = {version = "^0.15.1", optional = true}
pytest-cov = {version = "^3.0", optional = true}
pytest-benchmark = {version = "^3.4.1", optional = true}
types-orjson = {version = "^0.1.1", optional = true}
factory-boy = {version = "^3.2.1", optional=true}
sqlparse = "^0.4.2"

[tool.poetry.extras]
asyncpg = ["asyncpg", "orjson"]
psycopg = ["psycopg", "orjson"]
docs = ["mkdocs", "mkdocs-material", "mkdocs-awesome-pages-plugin"]
lint = ["mypy", "black", "flake8", "types-orjson"]
tests = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-benchmark",
    "asyncpg",
    "orjson",
    "psycopg",
    "aiosqlite",
    "factory-boy",
]

[tool.poetry.dev-dependencies]
# Docs
mkdocs = "^1.2.2"
mkdocs-material = "^8.0"
mkdocs-awesome-pages-plugin = "^2.5.0"
# Lint
mypy = "^0.910"
black = "^21.7b0"
flake8 = "^3.9.2"
# Test
pytest = "^6.2.4"
pytest-asyncio = "^0.15.1"
pytest-cov = "^3.0"
pytest-benchmark = "^3.4.1"
pre-commit = "^2.13.0"
asyncpg = "^0.25.0"
orjson = "^3.6.0"
psycopg = { extras = ["binary", "pool"], version = "^3.0"}
ptpython = "^3.0.20"
factory-boy = "^3.2.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
